+++ examples/basic/main.tf
@@ -1,6 +1,13 @@
 provider "aws" {
   region = "eu-central-1"
 
+  default_tags {
+    tags = {
+      Terraform = "Yes"
+      Test      = "Yes"
+    }
+  }
+
   # Make it faster by skipping the following
   skip_get_ec2_platforms      = true
   skip_metadata_api_check     = true
@@ -9,35 +16,30 @@ provider "aws" {
   skip_requesting_account_id  = true
 }
 
+data "aws_default_tags" "current" {}
 
 module "basic" {
   source = "../../"
-  # direction - (required) is a type of string
-  direction = "INBOUND"
-  # name - (optional) is a type of string
-  name = "basic"
-  # security_group_ids - (required) is a type of set of string
-  security_group_ids = ["sg-12345678901234567"]
-  # tags - (optional) is a type of map of string
-  tags = {
-    test      = "yes"
-    terraform = "yes"
-  }
 
-  ip_address = [{
-    ip        = null
-    ip_id     = null
-    subnet_id = "subnet-12345678901234567"
+  direction          = "INBOUND"
+  name               = "basic"
+  security_group_ids = ["sg-12345678901234567"]
+  tags = merge(
+    data.aws_default_tags.current.tags,
+    {
+      Name = "R53-INBOUND"
+    }
+  )
+
+  ip_address = [
+    {
+      ip        = null
+      ip_id     = null
+      subnet_id = "subnet-12345678901234567"
     },
     {
       subnet_id = "subnet-12345678907654321"
     },
   ]
-
-  timeouts = [{
-    create = null
-    delete = null
-    update = null
-  }]
 
 }