+++ variables.tf
@@ -31,14 +31,3 @@ variable "ip_address" {
   ))
 }
 
-variable "timeouts" {
-  description = "nested block: NestingSingle, min items: 0, max items: 0"
-  type = set(object(
-    {
-      create = string
-      delete = string
-      update = string
-    }
-  ))
-  default = []
-}