+++ CHANGELOG.md
@@ -7,20 +7,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2022-02-10
+
+### Added 2.0.0
+
+- CONTRIBUTING
+- CODE_OF_CONDUCT
+- CODEOWNERS
+- LICENSE
+- `aws_default_tags`
+
+### Changed 2.0.0
+
+- cleanup
+- refactor
+- update terraform workflow for github actions
+- remove tflint
+- Best Practices for Provider Versions
+- improve outputs
+
 ## [1.2.0] - 2021-04-12
 
-### Added
+### Added 1.2.0
 
 - `ip_addresses` output
 
 ## [1.1.0] - 2021-04-12
 
-### Changed
+### Changed 1.0.0
 
 - Experimental: Optional Object Type Attributes by [@mhmdio](https://github.com/mhmdio).
 
 ## [1.0.0] - 2021-04-12
 
-### Added
+### Added 1.0.0
 
 - initial commit by [@mhmdio](https://github.com/mhmdio).